+++ frontend/portfolio-website/src/pages/tags/[tag].js
@@ -48,7 +48,8 @@ function TagsPage() {
             <GenericTemplate>
                 <HeroBanner
                     pageTitle={main.heroBanner.content[0].pageTitle}
-                    subheading={main.heroBanner.content[1].subheading}
+                    // tags do not get rendered correctly in the subheading
+                    // subheading={main.heroBanner.content[1].subheading}
                 />
                 {main.section__projects.content[1].children[0].content.list.length > 0 && (
                     <Section headingTitle={main.section__projects.content[0].title.main}>
@@ -140,10 +141,10 @@ export async function getStaticProps({ params: { tag, name } }) {
     content.excerpt = content.excerpt.replace('{0}', tagName);
     content.tags.push(tagName);
     content.main.heroBanner.content[0].pageTitle = `#${tagName}`;
-    content.main.heroBanner.content[1].subheading = content.main.heroBanner.content[1].subheading.replace(
-        '{0}',
-        tagName
-    );
+    // content.main.heroBanner.content[1].subheading = content.main.heroBanner.content[1].subheading.replace(
+    //     '{0}',
+    //     tagName
+    // );
     content.url = content.url.replace('{0}', tag);
 
     const projectsFrontmatter = projects.map((project) => project.content.frontmatter);