+++ wiki/CHANGELOG.md
@@ -1,3 +1,32 @@
+# [1.17.0](https://github.com/paulAlexSerban/prj--personal-portfolio/compare/v1.16.4...v1.17.0) (2024-01-30)
+
+
+### Features
+
+* remove resume from footer ([7162834](https://github.com/paulAlexSerban/prj--personal-portfolio/commit/71628342b0a8c470d86b1ba4ae378da480901c3c))
+
+## [1.16.4](https://github.com/paulAlexSerban/prj--personal-portfolio/compare/v1.16.3...v1.16.4) (2024-01-25)
+
+
+### Bug Fixes
+
+* **tags still appear:** another tryout ([ee2975c](https://github.com/paulAlexSerban/prj--personal-portfolio/commit/ee2975c35203e2cc76eb5fc02e292a3330e17eaf))
+
+## [1.16.3](https://github.com/paulAlexSerban/prj--personal-portfolio/compare/v1.16.2...v1.16.3) (2024-01-24)
+
+
+### Bug Fixes
+
+* **fix tag in subheading issue:** fixed ([6f25434](https://github.com/paulAlexSerban/prj--personal-portfolio/commit/6f25434b79fa192342bacdd248cf2308bf37999e))
+* **tag subheading:** removed as it does not work propoerly ([1a95949](https://github.com/paulAlexSerban/prj--personal-portfolio/commit/1a9594999f3f51b4fe07c7b565baba041314f8db))
+
+## [1.16.2](https://github.com/paulAlexSerban/prj--personal-portfolio/compare/v1.16.1...v1.16.2) (2023-10-17)
+
+
+### Bug Fixes
+
+* remove user select css rule ([e280f91](https://github.com/paulAlexSerban/prj--personal-portfolio/commit/e280f91d66cbe264176f2353eb0420eaa4f18e91))
+
 ## [1.16.1](https://github.com/paulAlexSerban/prj--personal-portfolio/compare/v1.16.0...v1.16.1) (2023-09-28)